removed core.ts
