+++ src/app.native.ts
@@ -7,7 +7,6 @@ export * from "./app/config.js"
 export * from "./app/client.js"
 export * from "./app/logger.js"
 export * from "./app/slash.js"
-export * from "./app/core.js"
 export * from "./app/util.js"
 export * from "./config.js"
 
@@ -18,5 +17,4 @@ export * as listener from "./app/listener.js"
 export * as command from "./app/command.js"
 export * as logger from "./app/logger.js"
 export * as slash from "./app/slash.js"
-export * as core from "./app/core.js"
 export * as util from "./app/util.js"
